Revert to custom icons.
In da4adfbc and a505fdef the custom icons were replaced by KDE icons.
As a general rule it is better to use KDE icons, but not in this case.
Even if the appearance of the icons in the default theme suites, their
semantic does not. For example, a "dialog apply" does not have the same
meaning as a "successful test", even if they look similar in the default
theme.