Revert showing elapsed time for each test.
Since commit 5ed99701 the information caption showed the time spent
executing each test (or at least it was suppossed to according to the
log, but it didn't really worked). I'm reverting that behaviour to only
show the total elapsed time. Unit tests are usually very quick, so
updating the caption after finishing each test with the time spent
executing it is virtually unreadable. The total time, as it was before,
is a better approach.